1. ScrollContext added, moved all eventHandlers to it. 2. Create post button moved to App from Home component.
