+++ frontend/src/App.jsx
@@ -1,9 +1,10 @@
-import React, { useContext, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import "./App.css";
 import { Link, Outlet, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import axios from "axios";
-import UserContext from "./context/UserContext";
+import { ScrollContext, UserContext } from "./context/Contexts";
+import CreatePost from "./components/CreatePost";
 
 const App = () => {
   const categories = [
@@ -46,7 +47,14 @@ const App = () => {
     setInputValue("");
   };
 
-  const { isScrolledPast } = useContext(UserContext);
+  const { isVisible, isScrolledPast } = useContext(ScrollContext);
+  const [displayCreatePost, setDisplayCreatePost] = useState(false);
+  const [showPostedMessage, setShowPostedMessage] = useState(false);
+
+  const handleCreate = () => {
+    console.log("creating new post");
+    setDisplayCreatePost(true);
+  };
 
   return (
     <>
@@ -110,6 +118,18 @@ const App = () => {
           </HeaderAndCategory>
 
           <Outlet />
+
+          <Create onClick={handleCreate} $isvisible={isVisible}>
+            <i className="fa-regular fa-plus"></i>
+          </Create>
+
+          <CreatePostWrapper $display={displayCreatePost}>
+            <CreatePost
+              setShowPostedMessage={setShowPostedMessage}
+              setDisplayCreatePost={setDisplayCreatePost}
+            />
+          </CreatePostWrapper>
+          <PostedMsg $display={showPostedMessage}>Posted</PostedMsg>
 
           {/* <BlankSpace></BlankSpace> */}
           {/* Temporary code */}
@@ -269,4 +289,60 @@ const InputBox = styled.div`
 
 const BlankSpace = styled.div`
   height: 100vh;
+`;
+
+const Create = styled.div`
+  position: fixed;
+  width: 60px;
+  height: 60px;
+  bottom: ${(props) => (props.$isvisible == 1 ? "0" : "-200px")};
+  right: calc(50vw - 300px);
+  background-color: #9b1f24;
+  color: white;
+  font-size: 1.7rem;
+  margin: 0 30px 30px 0;
+  border-radius: 50%;
+  transition: bottom 0.5s ease-out;
+  cursor: pointer;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  @media (max-width: 600px) {
+    right: 10px; /* Adjust for mobile screens */
+    width: 85px;
+    height: 85px;
+    margin: 0 50px 50px 0;
+  }
+`;
+
+const CreatePostWrapper = styled.div`
+  width: 600px;
+  height: 100%;
+  background-color: rgba(18, 8, 8, 0.1);
+  /* border: 1px solid rgba(255,255,255,0.1); */
+  backdrop-filter: blur(2px);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+  position: fixed;
+  top: 0;
+  display: ${(props) => (props.$display ? "block" : "none")};
+`;
+
+const PostedMsg = styled.div`
+  position: fixed;
+  bottom: ${(props) => (props.$display ? "20px" : "-100px")};
+  opacity: ${(props) => (props.$display ? 1 : 0)};
+  left: 50%;
+  transform: translate(-50%, -50%);
+  transition: bottom 1s ease-in-out, opacity 1s ease-in-out;
+
+  background-color: #303030;
+  color: #13cb35;
+  padding: 10px 20px;
+  border-radius: 5px;
+  font-size: 1.5rem;
 `;
