+++ frontend/src/components/ProfilePage.jsx
@@ -220,10 +220,4 @@ const PostWrapper = styled.div`
       object-fit: cover;
     }
   }
-
-  /* img{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  } */
 `;
