+++ frontend/src/context/Contexts.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const UserContext = React.createContext();
+const ScrollContext = React.createContext();
+
+export { ScrollContext, UserContext };
