--- frontend/src/context/UserContext.jsx
@@ -1,3 +1,0 @@
-import React from "react";
-const UserContext = React.createContext();
-export default UserContext;
