+++ frontend/src/main.jsx
@@ -1,15 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
-import UserContextProvider from './context/UserContextProvider.jsx'
-import Routers from './components/Routers.jsx'
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App.jsx";
+import "./index.css";
+import UserContextProvider from "./context/UserContextProvider.jsx";
+import Routers from "./components/Routers.jsx";
+import ScrollContextProvider from "./context/ScrollContextProvider.jsx";
 
-ReactDOM.createRoot(document.getElementById('root')).render(
+ReactDOM.createRoot(document.getElementById("root")).render(
   // <React.StrictMode>
-  <UserContextProvider>
-    {/* <App /> */}
-    <Routers />
-  </UserContextProvider>
+  <ScrollContextProvider>
+    <UserContextProvider>
+      {/* <App /> */}
+      <Routers />
+    </UserContextProvider>
+  </ScrollContextProvider>
   // </React.StrictMode>,
-)
+);
